Added functionality and fixed file identification
Script now creates it's own folder if it does not already exist, and organizes files created by the app.
Fixed an issue with identification of 64 bit install file for windows. The script original read the architecture type and compared to the install file name. So the file being looked for would need to contain x86_64, whereas the windows 64 install file contained x64. Script now successfully locates the file for downloading.